Add icons to category shop

+++ src/components/UpdateProduct.jsx
@@ -81,7 +81,6 @@ function Options({ item }) {
           initialQuantity === 0 ? parseInt(`${quantity}`, 10) : initialQuantity,
         quantity: parseInt(`${quantity}`, 10),
         cost: parseFloat(`${cost}`),
-        isUsed: false,
       });
 
       if (response.status !== 'success') {
@@ -183,7 +182,7 @@ function Options({ item }) {
 }
 
 function Actions() {
-  const { api, item, goBack, onUpdate, onDelete, isOnline } =
+  const { api, item, goBack, onDelete, isOnline } =
     useContext(ComponentContext);
   const { id } = item;
   const onClickUsed = async () => {
@@ -191,15 +190,11 @@ function Actions() {
     try {
       await api.updateItem({
         itemId: id,
-        expirationDate: item.expirationDate,
-        initialQuantity: item.initialQuantity,
-        quantity: item.quantity,
-        cost: item.cost,
         isUsed: true,
       });
+      item.isExpired = false;
       item.isUsed = true;
-      item.is_used = true;
-      onUpdate(item);
+      onDelete(item);
     } catch (e) {
       alert(e.message);
       return;
@@ -212,7 +207,36 @@ function Actions() {
           quantity: item.quantity,
           cost: item.cost,
         });
-        onDelete(item);
+      } catch (e) {
+        alert(e.message);
+        return;
+      }
+    }
+
+    goBack();
+  };
+  const onClickIsExpired = async () => {
+    if (!confirm('Are you sure you want to set the item as expired?')) return;
+    try {
+      await api.updateItem({
+        itemId: id,
+        isExpired: true,
+      });
+      item.isUsed = false;
+      item.isExpired = true;
+      onDelete(item);
+    } catch (e) {
+      alert(e.message);
+      return;
+    }
+
+    if (confirm('Do you want to add to the shopping list?')) {
+      try {
+        await api.addItemToShopping({
+          productId: item.product.id,
+          quantity: item.quantity,
+          cost: item.cost,
+        });
       } catch (e) {
         alert(e.message);
         return;
@@ -250,6 +274,16 @@ function Actions() {
           Set Item as Used
         </button>
       ) : null}
+      {moment(item.expirationDate).isAfter(moment(new Date())) ? null : (
+        <button
+          type="button"
+          className={styles.expired}
+          disabled={!isOnline}
+          onClick={onClickIsExpired}
+        >
+          Set Item as Expired
+        </button>
+      )}
 
       <button
         type="button"
